+++ modules/src/compression/lossless/arencode2.c
@@ -600,7 +600,7 @@ arencode2(int *Print, long int *Size, int *NSymb, long int *Cap_Histo, int *Pred
     ncolo = 1;
     nrowo = ncodewords;
     if (nrowo > MAX_SIZEO) {
-      if ((int) sqrt((double) nrowo) + 1 > MAX_SIZEO)
+      if (sqrt((double) nrowo) + 1 > MAX_SIZEO)
 	mwerror(FATAL, 2, "Number of codewords is too large!\n");
       i = 2;
       while (nrowo / i > MAX_SIZEO)
